+++ src/utils/index.ts
@@ -2,3 +2,5 @@ export * from './with-request-timeout';
 export * from './truncate';
 export * from './get-package-version';
 export * from './check-for-update';
+export * from './read-response-text';
+export * from './validate-https-url';